Moved dispatch_key_event method from script_task.rs to document.rs, fixes #4982
